requirements/requirements.txt
mock==2.0.0
pytz==2018.3
requests==2.20.0
testfixtures==6.0.1
Django==2.0.9
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt=1.11.0
inquirer==2.2.0

gunicorn==19.9.0

